Fix enabling/disabling RM, added power setting
When first enabled, the S2 paths are added and both OMBC and No control are offered
After this, disabling the S2 function will change the allowed control types
to only "No control" to signal HEMS that the RM is disabled.
Enabling it again will allow the OMBC control type as well.

Enable the S2 RM by writing 6 to /SwitchableOutput/x/Settings/Function

Also added a power setting path (/Devices/x/PowerSetting) which is the nominal appliance power configured by the user.
The value of this setting is used in the 'on' operational mode of the OMBC system description.
Changing it will cause the systemdescription to be sent again.