Ignore calculated hue when sat is very low
For very low saturation, the hue calculation will be inaccurate.
If this hue value is used, it will cause small jumps in the hue when changing the saturation.
This resolves that issue.

Note that this may cause a change in hue to be ignored when
it comes from an external source and the sat is below this threshold.
However, the shelly app does not allow the hue and sat to be changed separately,
so this is not expected to happen normally.