Prevent updating /Dimming and /LightControls when an update is in progress
When sending multiple update commands (by dragging the slider in the UI),
prevent that the /Dimming and /LightControls paths are updated when a command to the shelly is still in progress.
This fixes the behavior where for example the brightness slider jumps back and forth
when the brightness is controlled from another UI instance.