Support white mode of RGBW devices.
In white mode (also called "Lights x 4"), a shelly RGBW PM device is used for 4 separate dimming channels.
It will show up as a device with 4 dimming channels.

This configuration change will reboot the device, so dbus-shelly will lose the connection and will start the reconnection mechanism.
After reconnecting, the shelly device will have a different number of channels and/or different capabilities.
If this is the case, all currently enabled services of that shelly device are disabled and a rediscovery is done
so it will show up in the integrations menu with the correct number of channels.